Create upload_togdrive
This script contains functions, that automate the downloading of data from web URLs and uploading them to a specific folder in Google Drive. 
It is needed, because of 50 mb filesize limit of gdrive. File is split on the server, uploaded in parts to gdrive, and assembled once again into one file. 
This code is used to synchronize data from the downloaded files to an Excel sheet stored in Google Drive.